chinta-auth/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
authlib>=1.3.0
httpx>=0.26.0
pyyaml>=6.0.1
